pyproject.toml
[project]
name = "kosokolovsky_telegram_bot"
version = "0.0.1"
description = "Telegram bot by Kosokolovsky"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "boto3",
    "requests",
    "python-telegram-bot",
    "python-dotenv",
    "configparser"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
